feat(gateway): make the help show if no args provided, set a default file to load for config
